use regular int

+++ domain/character.go
@@ -14,15 +14,15 @@ type Character struct {
 
 type CharacterUsecase interface {
 	Fetch(ctx context.Context, page int) ([]int, error)
-	GetByID(ctx context.Context, id int64) (Character, error)
+	GetByID(ctx context.Context, id int) (Character, error)
 }
 
 type CharacterReadRepository interface {
         Fetch(ctx context.Context, page int) ([]int, error)
-        GetByID(ctx context.Context, id int64) (Character, error)
+        GetByID(ctx context.Context, id int) (Character, error)
 }
 
 type CharacterWriteRepository interface {
         StoreByPage(ctx context.Context, page int) error
-        StoreByID(ctx context.Context, id int64) error
+        StoreByID(ctx context.Context, id int) error
 }
